AG-12847 - extract agg column name logic
