+++ community-modules/core/src/interfaces/iAggColumnNameService.ts
@@ -1,0 +1,5 @@
+import type { AgColumn } from '../entities/agColumn';
+
+export interface IAggColumnNameService {
+    getHeaderName(column: AgColumn, headerName: string | null): string | null;
+}
